{"schema":"decodex/commit/1","summary":"Split orchestrator status history projection","authority":"XY-1115"}
